requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
dill==0.2.8.2
genie==3.1.0
genie.abstract==3.1.0
genie.conf==3.1.0
genie.examples==3.1.1
genie.harness==3.1.2
genie.libs.conf==3.1.1
genie.libs.filetransferutils==3.1.0
genie.libs.ops==3.1.2
genie.libs.parser==3.1.5
genie.libs.sdk==3.1.4
genie.libs.telemetry==3.1.0
genie.metaparser==3.1.0
genie.ops==3.1.1
genie.parsergen==3.1.1
genie.predcore==3.1.0
genie.telemetry==3.1.0
genie.utils==3.1.0
httplib2==0.12.0
idna==2.7
ipyats==0.2.0
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
Jinja2==2.10
jsonpickle==1.0
junit-xml==1.8
MarkupSafe==1.1.0
netaddr==0.7.19
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psutil==5.4.8
ptyprocess==0.6.0
pyats==5.0.1
pyats.aereport==5.0.1
pyats.aetest==5.0.0
pyats.async==5.0.0
pyats.connections==5.0.0
pyats.datastructures==5.0.0
pyats.easypy==5.0.0
pyats.examples==5.0.0
pyats.kleenex==5.0.0
pyats.log==5.0.0
pyats.results==5.0.0
pyats.tcl==5.0.0
pyats.templates==5.0.0
pyats.topology==5.0.0
pyats.utils==5.0.1
Pygments==2.2.0
PyYAML==3.13
requests==2.20.1
setproctitle==1.1.10
six==1.11.0
traitlets==4.3.2
unicon==3.4.0
urllib3==1.24.1
wcwidth==0.1.7
xmltodict==0.11.0
